+++ packages/@expo/cli/CHANGELOG.md
@@ -12,6 +12,8 @@
 
 - Drop outdated React Native resolver patch. ([#29214](https://github.com/expo/expo/pull/29214) by [@EvanBacon](https://github.com/EvanBacon))
 
+- Use Metro instance directly for server rendering. ([#28552](https://github.com/expo/expo/pull/28552) by [@EvanBacon](https://github.com/EvanBacon))
+
 ## 0.18.13 — 2024-05-16
 
 ### 🐛 Bug fixes